// packages/data-explorer/src/types/ResourceCollection.ts
// tslint:disable
/**
 * MOLGENIS data & metadata API
 * RESTful API to create/read/update/delete data and metadata.
 *
 * The version of the OpenAPI document: 3.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Links } from '@/types/Links.ts'
import { Page } from '@/types/Page'
import { Resource } from '@/types/Resource'

/**
 *
 * @export
 * @interface ResourceCollection
 */
export interface ResourceCollection {
    /**
     *
     * @type {Links}
     * @memberof ResourceCollection
     */
    links: Links;
    /**
     *
     * @type {Array<Resource>}
     * @memberof ResourceCollection
     */
    items: Array<Resource>;
    /**
     *
     * @type {Page}
     * @memberof ResourceCollection
     */
    page: Page;
}