// packages/data-explorer/src/types/Resource.ts
// tslint:disable
/**
 * MOLGENIS data & metadata API
 * RESTful API to create/read/update/delete data and metadata.
 *
 * The version of the OpenAPI document: 3.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { Links } from '@/types/Links.ts'
import { ResourceCollection } from './ResourceCollection'

/**
 *
 * @export
 * @interface Resource
 */
export interface Resource {
    /**
     *
     * @type {Links}
     * @memberof Resource
     */
    links: Links;
    /**
     *
     * @type {object}
     * @memberof Resource
     */
    data?: Record<string, string | number | boolean | ResourceCollection | Resource>;
}